Added some polish and replaced a similar template
