+++ templates/Laser-Grunt_gamedata.json
@@ -11,7 +11,16 @@
         "botgrunt"
     ],
     "ai_job": [
-        "wait"
+        "wander"
+    ],
+    "AI_Aggression": [
+        100.0
+    ],
+    "AI_Intelligence": [
+        100.0
+    ],
+    "AI_Courage": [
+        100.0
     ],
     "NPCWeapon0": [
         "laser"
@@ -19,35 +28,11 @@
     "NPCWeapon1": [
         "grenade"
     ],
-    "ai_intelligence": [
-        100.0
-    ],
-    "ai_courage": [
-        100.0
-    ],
-    "AI_Aggression": [
-        100.0
-    ],
-    "AI_ODDSOF_IGNORING_NEARBY_GRENADES": [
-        0.0
-    ],
-    "AI_PREFERSTOPANDSHOOT": [
-        "false"
-    ],
-    "AI_GRENADE_USE": [
-        "whenever"
-    ],
-    "AI_BONUS_EYE_SCAN_DIST_FOR_PLAYERS_ALERT": [
-        500.0
-    ],
-    "AI_ATTACK_NOTIFY_RADIUS": [
-        175.0
-    ],
-    "goodie1": [
-        "goodiebag=1",
+    "goodiebag": [
+        1.0,
         "one"
     ],
-    "goodie1__1": [
+    "goodie1": [
         "washer",
         5.0,
         5.0,
